feat(ui): make server entry scale slider fully continuous (float 1.0-4.0)

+++ settings/src/main/kotlin/com/spoiligaming/explorer/settings/model/MultiplayerSettings.kt
@@ -59,7 +59,7 @@ data class MultiplayerSettings(
     @SerialName("server_query_method")
     val serverQueryMethod: ServerQueryMethod = ServerQueryMethod.McServerPing,
     @SerialName("entry_size_scale")
-    val serverEntryScale: Int = 1,
+    val serverEntryScale: Float = 1.0f,
     @SerialName("drag_shake_intensity_deg")
     val dragShakeIntensityDegrees: Int = 1,
     @SerialName("action_bar_orientation")
